pyproject.toml
[project]
name = "gif"
version = "0.1.0"
description = ""
authors = [
    {name = "alwaysmavs",email = "beingswu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "imageio (>=2.37.0,<3.0.0)"
]

[tool.poetry]
package-mode = false
packages = [{include = "gif", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
